esta mejor la forma en que se lee el archivo en el parser por algun motivo no funcionaba en gentoo y ahora anda :)

+++ common/map.h
@@ -9,7 +9,7 @@ struct rawbrush{
 };
 
 struct brushlist{
-  unsigned long int used;//para crear los arboles
+        int used;//para crear los arboles
 	struct rawbrush *rbsh;
 	struct brush	*bsh;
 	struct brushlist *next;
